requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
bcrypt==4.1.2
boto3==1.34.72
botocore==1.34.72
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cramjam==2.8.1
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==4.1.13
django-cors-headers==4.3.1
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
djoser==2.2.3
dnspython==2.6.1
email_validator==2.1.1
gunicorn==22.0.0
idna==3.6
jmespath==1.0.1
mango-jwt==1.3.2
mongoengine==0.28.2
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
passlib==1.7.4
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==3.12.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-snappy==0.7.1
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.1
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.2.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
winkerberos==0.10.0
